Move generator tests to own module
